backend/requirements-prod.txt
# 仅生产依赖，用于 Docker 构建（不包含 pytest 等）
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=43.0.0
pydantic[email]>=2.9.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
jieba>=0.42.1
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
openai>=1.50.0
httpx>=0.27.0
